+++ README.md
@@ -9,17 +9,19 @@ cargo add duners
 ```
 
 ```rust
-use std::env;
+use chrono::{DateTime, Utc};
 use dotenv::dotenv;
-use duners::client::{DuneClient, DuneRequestError};
+use duners::{client::DuneClient, dateutil::datetime_from_str};
 use serde::Deserialize;
+use std::env;
 
 // User must declare the expected query return fields and types!
 #[derive(Deserialize, Debug, PartialEq)]
 struct ResultStruct {
     text_field: String,
-    number_field: String,
-    date_field: String,
+    number_field: f64,
+    #[serde(deserialize_with = "datetime_from_str")]
+    date_field: DateTime<Utc>,
     list_field: String,
 }
