+++ src/lib.rs
@@ -1,9 +1,10 @@
 /// DuneClient structure and all API route implementations.
 pub mod client;
+/// Utility Methods (primarily for date parsing)
+pub mod dateutil;
 /// DuneRequestError (encapsulating all errors that could arise within network requests and result parsing)
 pub mod error;
 /// Content related to Query Parameters.
 pub mod parameters;
 /// Data models representing response types for all client methods.
 pub mod response;
-mod util;
